Add mocked test auth user to reuse in tests.
