+++ coral/src/domain/auth-user/auth-user-test-helper.ts
@@ -1,0 +1,11 @@
+import { AuthUser } from "src/domain/auth-user/auth-user-types";
+
+const testAuthUser: AuthUser = {
+  canSwitchTeams: "false",
+  teamId: "1234567",
+  teamname: "awesome-bunch-of-people",
+  username: "i-am-test-user",
+  userrole: "USER",
+};
+
+export { testAuthUser };
